+++ src/rx-collection.ts
@@ -508,10 +508,6 @@ export class RxCollectionBase<
 
         return this.findOne(primary).exec()
             .then((existing: RxDocument<RxDocumentType, OrmMethods> | null) => {
-
-                console.log('existing:');
-                console.dir(existing);
-
                 if (existing && !existing.deleted) {
                     useJson._rev = (existing as any)['_rev'];
                     return existing.atomicUpdate(() => useJson as any)